+++ src/main/java/dev/_2lstudios/squidgame/arena/games/G6GlassesGame.java
@@ -1,26 +1,38 @@
 package dev._2lstudios.squidgame.arena.games;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import org.bukkit.Material;
 import org.bukkit.World;
+import org.bukkit.block.Block;
 
 import dev._2lstudios.jelly.math.Cuboid;
 import dev._2lstudios.jelly.math.Vector3;
+import dev._2lstudios.jelly.utils.BooleanUtils;
 import dev._2lstudios.squidgame.arena.Arena;
 
 public class G6GlassesGame extends ArenaGameBase {
 
     private Cuboid glassZone;
+    private List<Block> fakeBlocks;
 
     public G6GlassesGame(final Arena arena, final int durationTime) {
         super("§bGlasses", "sixth", durationTime, arena);
+
+        this.fakeBlocks = new ArrayList<>();
     }
 
-    public Cuboid getGlassZone() {
+    private Cuboid getGlassZone() {
         if (this.glassZone == null) {
             this.glassZone = this.getArena().getConfig().getCuboid("games.sixth.glass_zone");
         }
 
         return this.glassZone;
+    }
+
+    public boolean isFakeBlock(final Block block) {
+        return this.fakeBlocks.contains(block);
     }
 
     @Override
@@ -68,6 +80,9 @@ public class G6GlassesGame extends ArenaGameBase {
 
         // Por cada grupo de plataforma (+ 1) iterar:
         for (int i = 0; i <= platformGroups; i++) {
+            // Is first pair item a fake block?
+            boolean isFirstFake = BooleanUtils.randomBoolean();
+
             // En caso que la coordenada Z deba de usarse como un indice:
             if (useZAsIndex) {
                 // Por cada posición relativa X de la plataforma
@@ -80,9 +95,18 @@ public class G6GlassesGame extends ArenaGameBase {
                         int z = shouldIncreaseIndex ? blockIndex + zPadding : blockIndex - zPadding;
 
                         // Generar bloque en las coordenadas dadas X Y Z
-                        world.getBlockAt(x, yStart, z).setType(Material.BEDROCK);
+                        final Block firstRowBlock = world.getBlockAt(x, yStart, z);
+                        firstRowBlock.setType(Material.GLASS);
+
                         // Generar bloque en la misma posicion que el de arriba pero con una separación
-                        world.getBlockAt(x + spaceXBetweenPlatforms + size, yStart, z).setType(Material.BEDROCK);
+                        final Block secondRowBlock = world.getBlockAt(x + spaceXBetweenPlatforms + size, yStart, z);
+                        secondRowBlock.setType(Material.GLASS);
+
+                        if (isFirstFake) {
+                            this.fakeBlocks.add(firstRowBlock);
+                        } else {
+                            this.fakeBlocks.add(secondRowBlock);
+                        }
                     }
                 }
 
@@ -108,6 +132,7 @@ public class G6GlassesGame extends ArenaGameBase {
             final int separation = spaceZBetweenPlatforms + size;
             blockIndex = shouldIncreaseIndex ? blockIndex + separation : blockIndex - separation;
         }
+
     }
 
     @Override